requirements.txt
aiogram>=3.4.1,<4.0.0
aiosqlite>=0.20.0,<1.0.0
APScheduler>=3.10.4,<4.0.0
python-dotenv>=1.0.1,<2.0.0
